compiler: Don't cast index expr to int before bounds check.
This fixes http://gcc.gnu.org/PR61866 : on a 32-bit system,
casting an int64 index to int drops the upper 32 bits of the
value, and thus can cause an out-of-range index to appear to
be in range.

This undoes part of change 1318:fa6e0c716dba
(https://golang.org/cl/104610044) and therefore
breaks http://gcc.gnu.org/PR61308 again.  I have a separate
patch for that (https://golang.org/cl/122020043).  In
addition to undoing part of that change, this patch adds code
to avoid a compiler crash.  This changes PR61308 from a
compiler crash to an incorrect error message.

R=cmang
CC=gofrontend-dev
https://golang.org/cl/119430043