escape: Analyze function values and conversions.
The analysis for variables defined as function literals did not
properly analyze the underlying function literal when it was converted
to a function type.  Also, when analyzing composite literals with
function literals as arguments, the connection between the composite
literal and the function literal was ignored.

LGTM=iant
R=iant
CC=gofrontend-dev
https://golang.org/cl/229530043