syscall:  RawSockaddr fix for ppc64, ppc64le
The struct RawSockaddr contains a field Data which
should be uint8 on ppc64 and ppc64le, but is declared
as int8 in gccgo.  This change adds a two new files
which contain the structure declaration for
RawSockaddr, one with the correct types for for ppc64
and ppc64le, and the other for non-ppc64 platforms.

Fixes golang/go#11469

Change-Id: I9964d4b04d63038a407580aa282129458201c2c1
Reviewed-on: https://go-review.googlesource.com/11946
Reviewed-by: Ian Lance Taylor <iant@golang.org>