compiler: Make empty interface types for vars during parse time.
When making the type for a variable with an empty interface type,
the parser makes an interface type with a NULL method set and relies
on later passes to correct this.

For sink variables, which are ignored in later passes, the interface
method table is never finalized and a compile time assertion is issued.
Instead, the initial type generated by the parser should be the empty
interface type.

Fixes golang/go#11579.

Change-Id: I479559f270ddf88afc7a33103c0f56afda195d94
Reviewed-on: https://go-review.googlesource.com/12049
Reviewed-by: Ian Lance Taylor <iant@golang.org>