compiler: Check the type in function declarations.
Function declarations don't create a block where the variables listed
in the parameter list are declared.  Because there are no variables
declared, the types of the parameter variables is unchecked, allowing
for invalid values to be used as the type.  This patch adds a special
case to the check_types pass for function declarations.

Fixes golang/go#11567.

Change-Id: I65569cb6f8b5aa8f64e5a28d0967036c8d12dda4
Reviewed-on: https://go-review.googlesource.com/12662
Reviewed-by: Ian Lance Taylor <iant@golang.org>