compiler: Flatten erroneous subtrees into errors.
Between the lowering and flattening passes of the compiler, there are
several passes that modify the lowered Go parse tree and as errors are
discovered, several nodes transform into error nodes.  However, for a
higher level node such as a construction expression, the erroneous
nodes in the subtrees might not propagate their error.  The flatten
phase for a node now looks for errors in the subtree and flattens the
node into an error node if any are found.

Fixes golang/go#11559, golang/go#11536, golang/go#11558.

Change-Id: I9a99d5a2f8ab396ca71787c2bd2af3b1f7ef625a
Reviewed-on: https://go-review.googlesource.com/13097
Reviewed-by: Ian Lance Taylor <iant@golang.org>