compiler: Accept untyped integral values as string/array indices.
When determining the type of an index for a string/array indexing
expression, the gofrontend would disallow floating-point and complex
values even if they were integral and throw an internal error.  This
patch changes gofrontend to use an integral type context when
determining the types of a string/array index.

Fixes golang/go#11545.

Change-Id: Ieebc3f66ed5181bdfc16ab94855349a6bb7f2403
Reviewed-on: https://go-review.googlesource.com/13796
Reviewed-by: Ian Lance Taylor <iant@golang.org>