compiler: Mark erroneous constants as invalid.
When the compiler failed to evaluate a numeric constant because
because it was erroneous, there was no way for parent nodes to
discover the error and lower themselves into error nodes.  This patch
now uses the NC_INVALID enumerator to mark numeric constants with a
known, reported error.

Fixes golang/go#11541.

Change-Id: I2cc78e9ecf9f946a94804cd96aa4dc51dd04c8fb
Reviewed-on: https://go-review.googlesource.com/13904
Reviewed-by: Ian Lance Taylor <iant@golang.org>