compiler: Issue channel type errors earlier.
When asking for the type of a receive operation, the compiler would
return an error type if the receive operator was being used on an
invalid channel type and the error would be reported in a later pass.
There are several ways that the type checking pass would not see
the original node and fail to issue the error.  This patch modifies
receive operations to give an error immediately once it is known that
the channel type is invalid.

Fixes golang/go#12323.

Change-Id: Ie3d9c33cf36fd05f49ed807ee52faaeb7d7cfdc2
Reviewed-on: https://go-review.googlesource.com/13987
Reviewed-by: Ian Lance Taylor <iant@golang.org>