compiler: Record each import as a distinct alias.
This patch introduces the Package_alias class which is a
finer representation of the different between a package and the
aliases it is imported under.  Each alias keeps track of the location
of its import statement and how many times that alias has been used.
This allows the gofrontend to report when a specific import has not
been used even if a symbol from the package has been used by another
import.

Fixes golang/go#12326.

Change-Id: I03bef25c5f9a32beded31a7b1a8140f7bea54ed2
Reviewed-on: https://go-review.googlesource.com/14259
Reviewed-by: Ian Lance Taylor <iant@golang.org>