compiler: Avoid knock-on errors with impossibly large types.
The gofrontend asks the backend compiler for the architecture
appropriate size of a given type.  For array types, it is possible
to construct a type too large to fit on the machine.  This patch does
two things: 1. When an impossibly large type is encountered, we mark
the type as erroneous and later calls to discover the size of that
type are short-circuited. 2. When generating the GC symbol data for
an impossibly large array of arrays, we avoid generating symbol data
as soon as we find an array that is too large to be expressed.

Fixes golang/go#12938.

Change-Id: If985b99a4ab54955be3dbd514fb94f3e271ef7b8
Reviewed-on: https://go-review.googlesource.com/16234
Reviewed-by: Ian Lance Taylor <iant@golang.org>