compiler: Don't crash on erroneous array return types.
Another issue with erroneous array types.  When an erroneous array
type is in a function's signature, particularly the return type, we
must guarantee that type is changed into an error type.  Otherwise,
any operations that work on arrays and slices will crash when applied
to the erroneous array return type.

Fixes golang/go#12939.

Change-Id: Ibd2445ad88b5dfd633b8bead0a880586d805bb16
Reviewed-on: https://go-review.googlesource.com/16235
Reviewed-by: Ian Lance Taylor <iant@golang.org>