compiler, runtime: change type hash/equal to Go funcs
Change the type descriptor hash and equal functions from C code pointers
to Go func values.  This permits them to be set to a Go function
closure.  This is in preparation for the Go 1.5, so that we can use a
closure for the hash/equal functions returned by the new reflect.ArrayOf
function.

Change-Id: I26dcdee4228eb0420d1b1ce1e423f814a5c1836f
Reviewed-on: https://go-review.googlesource.com/16485
Reviewed-by: Ian Lance Taylor <iant@golang.org>