cmd: Use correct install tool dir with gccgo
When using the go command built from gccgo to build and
install a go tool, use the value from runtime GCCGOTOOLDIR as
the install directory.

This also fixes the output from 'go tool' when used with the
gccgo-built go command, to only include the go tools and not
other binaries found in the same directory.

Change-Id: I5fea5639b76ed7ea43a1630fd971863ca6161af3
Reviewed-on: https://go-review.googlesource.com/16516
Reviewed-by: Ian Lance Taylor <iant@golang.org>