runtime: If no split stacks, allocate stacks using mmap on 64-bit systems.
When not using split stacks, libgo allocate large stacks for each
goroutine.  On a 64-bit system, libgo allocates a maximum of 128G for
the Go heap, and allocates 4M for each stack.  When the stacks are
allocated from the Go heap, the result is that a program can only create
32K goroutines, which is not enough for an active Go server.  This patch
changes libgo to allocate the stacks using mmap directly, rather than
allocating them out of the Go heap.  This change is only done for 64-bit
systems when not using split stacks.  When using split stacks, the
stacks are allocated using mmap directly anyhow.  On a 32-bit system,
there is no maximum size for the Go heap, or, rather, the maximum size
is the available address space anyhow.

Change-Id: I23c5b978bdd552bf26ddd4ae69b9e764d1927bcc
Reviewed-on: https://go-review.googlesource.com/16531
Reviewed-by: Ian Lance Taylor <iant@golang.org>