compiler: Traverse types of call expressions.
https://gcc.gnu.org/PR67968 provides a test case that causes a gccgo
crash on valid code.  The compiler failed to build the hash and equality
functions required for a type descriptor.  The descriptor is for an
unnamed type that is being returned by a function imported from a
different package.  The unnamed type is being implicitly converted to an
interface type by a return statement.  The fix is to ensure that the
type of a call expression is always traversed.

Test case sent out for the master testsuite as
https://golang.org/cl/16532 .

Change-Id: Id4e225fff6f0b57854073ee56e6fe51baaf0cf8a
Reviewed-on: https://go-review.googlesource.com/16533
Reviewed-by: Ian Lance Taylor <iant@golang.org>