libgo: Solaris portability fixes.
Only build net/hook_cloexec.go on GNU/Linux and FreeBSD, because those
are the only systems with accept4.

Add syscall/libcall_bsd.go to define sendfile for *BSD and Solaris.

Revert tcpsockopt_solaris.go back to the earlier version, so that it
works on Solaris 10.

Always pass the address of a Pid_t value to TIOCGPGRP and TIOCSPGRP.

Include <unistd.h> in runtime/go-varargs.c.

Change-Id: I9631613c2692058bd8aeae71860ce9a54e39529b
Reviewed-on: https://go-review.googlesource.com/16719
Reviewed-by: Ian Lance Taylor <iant@golang.org>