cmd/go: always use --whole-archive for gccgo packages
This is a backport of https://golang.org/cl/16775.

This is, in effect, what the gc toolchain does.  It fixes cases where Go
code refers to a C global variable; without this, if the global variable
was the only thing visible in the C code, the generated cgo file might
not get pulled in from the archive, leaving the Go variable
uninitialized.

This was reported against gccgo as https://gcc.gnu.org/PR68255 .

Change-Id: I8b8ac3761b6f5680d601d23f89b4784b4d869d30
Reviewed-on: https://go-review.googlesource.com/16778
Reviewed-by: Ian Lance Taylor <iant@golang.org>