reflect: Allocate space for FFI functions returning a zero-sized type.
The libffi library does not understand zero-sized types.  We represent
them as a struct with a single field of type void.  If such a type is
returned from a function, libffi will copy 1 byte of data.  Allocate
space for that byte, although we won't use it.

Fixes https://gcc.gnu.org/PR68496.

Change-Id: If4df92367191ee67b2bfbf90c3517ce5b7bfc9bb
Reviewed-on: https://go-review.googlesource.com/17175
Reviewed-by: Ian Lance Taylor <iant@golang.org>