libgo/testsuite: portable ps usage in gotest
gotest is using "ps" to list descendant sleep
processes in its timeout handling, grepping
the command name.

We are currently using the "cmd" ps output
formatter, which is non-portable. We should
use "comm" which is part of the POSIX
standard, and outputs only the command name
and not the arguments.

Fixes https://gcc.gnu.org/PR68980

Change-Id: I5292f842335d478bae8be8f0d86d31d501c3de36
Reviewed-on: https://go-review.googlesource.com/18426
Reviewed-by: Ian Lance Taylor <iant@golang.org>