runtime: For c-archive/c-shared, install signal handlers synchronously.
This is a port of https://golang.org/cl/18150 to the gccgo runtime.

The previous behaviour of installing the signal handlers in a separate
thread meant that Go initialization raced with non-Go initialization if
the non-Go initialization also wanted to install signal handlers.  Make
installing signal handlers synchronous so that the process-wide behavior
is predictable.

Change-Id: I4c7fe728ad8f77de3321474916a16875a2f3a7b2
Reviewed-on: https://go-review.googlesource.com/19494
Reviewed-by: Ian Lance Taylor <iant@golang.org>