gotest: multiple +build lines must all be true
The code that handled +build lines did not correctly require them to all
be true.  While looking into this I discovered that multiple +build
lines were mishandled in a different way, because the shell does not
preseve line breaks in backquoted data.  Look for the +build token to
tell us when we are switching from one +build line to another.

Change-Id: Ib91cba7ab928ddc3caaeb151c2560c1bdd2c8b75
Reviewed-on: https://go-review.googlesource.com/25460
Reviewed-by: Ian Lance Taylor <iant@golang.org>