runtime: treat zero-sized result value as void
Change the FFI interface to treat a call to a function that returns a
zero-sized result as a call to a function that returns void.

This is part of the fix for https://gcc.gnu.org/PR72814.  On 32-bit
SPARC systems, a call to a function that returns a non-zero-sized struct
is followed by an unimp instruction that describes the size of the
struct.  The function returns to the address after the unimp
instruction.  The libffi library can not represent a zero-sized struct,
so we wind up treating it as a 1-byte struct.  Thus in that case libffi
calls the function with an unimp instruction, but the function does not
adjust the return address.  The result is that the program attempts to
execute the unimp instruction, causing a crash.

This is part of a change that fixes the crash by treating all functions
that return zero bytes as functions that return void.

Change-Id: Ifb459a309340e4a41dfc2bb71b4ab19bbef2c8a9
Reviewed-on: https://go-review.googlesource.com/25585
Reviewed-by: Ian Lance Taylor <iant@golang.org>