compiler: revamp scheme for ordering calls to import init fcns.
Switch to a new method for determining the order in which import init
functions are invoked: build an init fcn dependence DAG and walk the DAG
to rewrite/adjust priorities to account for discrepancies introduced by
"go test".

This patch includes a change to the export data format generated
by gccgo. Older versions of gccgo will not be able to read object files
produced by a newer gccgo, but the new gcc will still be able to read
old object files.

Fixes golang/go#15738.

Change-Id: I0e5bec1b4595d8952b7a4d62545b0acdf84c6de2
Reviewed-on: https://go-review.googlesource.com/25301
Reviewed-by: Ian Lance Taylor <iant@golang.org>