runtime: use -fgo-c-header to build C header file
Use the new -fgo-c-header option to build a header file for the Go
runtime code in libgo/go/runtime, and use the new header file in the C
runtime code in libgo/runtime.  This will ensure that the Go code and C
code share the same data structures as we convert the runtime from C to
Go.

The new file libgo/go/runtime/runtime2.go is copied from the Go 1.7
release, and then edited to remove unnecessary data structures and
modify others for use with libgo.

The new file libgo/go/runtime/mcache.go is an initial version of the
same files in the Go 1.7 release, and will be replaced by the Go 1.7
file when we convert to the new memory allocator.

The new file libgo/go/runtime/type.go describes the gccgo version of the
reflection data structures, and replaces the Go 1.7 runtime file which
describes the gc version of those structures.

Using the new header file means changing a number of struct fields to
use Go naming conventions (that is, no underscores) and to rename
constants to have a leading underscore so that they are not exported
from the Go package.  These names were updated in the C code.

The C code was also changed to drop the thread-local variable m, as was
done some time ago in the gc sources.  Now the m field is always
accessed using g->m, where g is the single remaining thread-local
variable.  This in turn required some adjustments to set g->m correctly
in all cases.

Also pass the new -fgo-compiling-runtime option when compiling the
runtime package, although that option doesn't do anything yet.

Change-Id: I1d7cf75b89461ed5f890ef8f291e5a5e662cf516
Reviewed-on: https://go-review.googlesource.com/28051
Reviewed-by: Ian Lance Taylor <iant@golang.org>