runtime: align ucontext_t argument to 16 byte boundary
Some systems, such as ia64 and PPC, require that a ucontext_t pointer
passed to getcontext and friends be aligned to a 16-byte boundary.
Currently the ucontext_t fields in the g structure are defined in Go,
and Go has no way to ensure a 16-byte alignment for a struct field.
The fields are currently represented by an array of unsafe.Pointer.
Enforce the alignment by making the array larger, and picking an offset
into the array that is 16-byte aligned.

Change-Id: I96972d6dadf2bafb836e01228c262afaac8a5428
Reviewed-on: https://go-review.googlesource.com/28910
Reviewed-by: Ian Lance Taylor <iant@golang.org>