compiler: add abstraction layer for sha1 checksums.
Add new interface for the front end code to use when computing SHA1
checksums; the intent is to allow the different implementation in
different back ends.

No change in functionality for gccgo; this is an enabling change to
permit the front end to be used with other back ends (e.g. LLVM).

Change-Id: Ia9afcccc1fe4a1546cd2cdbde2e6fc79a6cea0bf
Reviewed-on: https://go-review.googlesource.com/28833
Reviewed-by: Ian Lance Taylor <iant@golang.org>