runtime/internal/sys: new package, API copied from Go 1.7
Copy over the Go 1.7 runtime/internal/sys package, but instead of having
separate files for each GOARCH and GOOS value, set the values in
configure.ac and write them out in Makefile.am.  Setting the values in
configure.ac should make it easier to add new processors.

Remove the automake GOARCH conditionals, which are no longer used.
Leave the GOOS conditionals for now, as they are used for the C runtime
package.

Change-Id: Ib958ba09c669fafc74b011f9fdb6bbcee46f0d2b
Reviewed-on: https://go-review.googlesource.com/29018
Reviewed-by: Ian Lance Taylor <iant@golang.org>